Refactor Signature component on review section
